Lists: Functional implementation, Scrollbar
Lists are now usable and are able to display a scrollbar with dynamic size. I wrote large parts of this while very tired so I might have to rethink the approach, as the current implementation is very clean for what it does, but it is very specialized and making a list go infinite or wrap around would require arbitrary value checks all over the place, so I won't be adding those until I come up with a better way to model UILists.

Currently scrolling lines that are too long isn't implemented, again because with the way this system is structured it would be rather hacky and end up like the first solution all over again