Fix Memory Leak
The List it uses wasn't being deleted when the object was destroyed resulting in a memory leak everytime the program menu was closed. This made me realize how stupid it is to initialize all the programs in the program list itself, so I'll have to think of a better way of doing that.